Refactor: Refactor Main Page Structure (#30)
[Refactor] Refactor Main Page Structure
